requirements.txt
pandas
numpy
argparse
dynaconf
pyspark==3.5.0
python-gitlab
requests
pyyaml
venv-pack